More type migration from double to float. Removed duplicate method from Food class

+++ src/data_access/EdamamApiAccess.java
@@ -2,6 +2,7 @@ package data_access;
 
 import entity.Recommend;
 import okhttp3.*;
+import org.json.JSONArray;
 import org.json.JSONObject;
 import java.io.IOException;
 import java.util.*;
@@ -49,7 +50,7 @@ public class EdamamApiAccess implements RecipeDataAccessInterface, RecommendData
         }
     }*/
 
-    public Integer getCalories(Food identifier) {
+    public float getCalories(Food identifier) {
         String foodName = identifier.getName();
         OkHttpClient client = new OkHttpClient();
